# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Data API Gateway
  description: "API Gateway multi-tenant (Maxtop, Pandora, Toplink) dengan dukungan Dual-DB (Postgres & MSSQL)."
  version: 1.0.0
...
servers:
  - url: /
    description: Current Server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: API Key Developer
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token yang didapatkan dari login
  parameters:
    TenantParam:
      name: tenant
      in: path
      required: true
      description: Nama Unit/Tenant (maxtop, pandora, toplink)
      schema:
        type: string
    DbSourceHeader:
      name: X-DB-Source
      in: header
      required: false
      description: Pilih mesin database (default menggunakan postgres)
      schema:
        type: string
        enum: [postgres, mssql]
        default: postgres
  schemas:
...
    Outbox:
      type: object
      properties:
        kode:
          type: integer
          format: int64
          example: 204300
        tgl_entri:
          type: string
          format: date-time
          example: "2026-09-06T10:00:00Z"
        penerima:
          type: string
          example: "08123456789"
        tipe_penerima:
          type: string
          example: "SMS"
        pesan:
          type: string
          example: "Saldo Anda: Rp100.000"
        status:
          type: integer
          example: 0
        tgl_status:
          type: string
          format: date-time
          nullable: true
          example: "2026-09-06T10:05:00Z"
        kode_inbox:
          type: integer
          format: int64
          nullable: true
        kode_transaksi:
          type: integer
          nullable: true
        kode_reseller:
          type: string
          nullable: true
        bebas_biaya:
          type: integer
          example: 0
        is_perintah:
          type: integer
          nullable: true
        kode_modul:
          type: integer
          nullable: true
        prioritas:
          type: integer
          nullable: true
        modul_proses:
          type: string
          nullable: true
        pengirim:
          type: string
          nullable: true
        kode_terminal:
          type: integer
          nullable: true
        ctr_kirim:
          type: integer
          nullable: true
    CreateOutboxRequest:
      type: object
      required:
        - penerima
        - tipe_penerima
        - pesan
        - status
        - bebas_biaya
      properties:
        penerima:
          type: string
        tipe_penerima:
          type: string
        pesan:
          type: string
        status:
          type: integer
        bebas_biaya:
          type: integer
        kode_inbox:
          type: integer
          format: int64
        kode_transaksi:
          type: integer
        kode_reseller:
          type: string
        is_perintah:
          type: integer
        kode_modul:
          type: integer
        prioritas:
          type: integer
        modul_proses:
          type: string
        pengirim:
          type: string
        kode_terminal:
          type: integer
        ctr_kirim:
          type: integer
    UpdateOutboxRequest:
      type: object
      properties:
        penerima:
          type: string
          nullable: true
        tipe_penerima:
          type: string
          nullable: true
        pesan:
          type: string
          nullable: true
        status:
          type: integer
          nullable: true
        bebas_biaya:
          type: integer
          nullable: true
        kode_inbox:
          type: integer
          format: int64
          nullable: true
        kode_transaksi:
          type: integer
          nullable: true
        kode_reseller:
          type: string
          nullable: true
        is_perintah:
          type: integer
          nullable: true
        kode_modul:
          type: integer
          nullable: true
        prioritas:
          type: integer
          nullable: true
        modul_proses:
          type: string
          nullable: true
        pengirim:
          type: string
          nullable: true
        kode_terminal:
          type: integer
          nullable: true
        ctr_kirim:
          type: integer
          nullable: true
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: sukses
        data:
          type: object
          additionalProperties: true
...
    CursorPaginationMeta:
      type: object
      properties:
        has_next_page:
          type: boolean
        has_prev_page:
          type: boolean
        next_cursor:
          type: integer
          format: int64
...
    OutboxResponse:
      type: object
      properties:
        trace_id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/Outbox"
        meta:
          $ref: "#/components/schemas/CursorPaginationMeta"
...
paths:
...
  /api/v1/{tenant}/outbox:
    get:
      summary: Get Outbox (12 Opsi Filter)
      description: "Mengambil daftar pesan Outbox"
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      tags:
        - outbox
      parameters:
        - $ref: "#/components/parameters/TenantParam"
        - $ref: "#/components/parameters/DbSourceHeader"
        - name: startDate
          in: query
          description: Tanggal entri mulai (Format YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Tanggal entri akhir (Format YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: pageSize
          in: query
          description: Jumlah data per halaman (default 10, maksimal 500)
          schema:
            type: integer
        - name: limit
          in: query
          description: Total data maksimal yang boleh di-deliver selama pagination (opsional, maksimum 100.000)
          schema:
            type: integer
        - name: reseller
          in: query
          description: Filter kode reseller (kode persis / dropdown)
          schema:
            type: string
        - name: penerima
          in: query
          description: Filter penerima (pencocokan sebagian / ILIKE)
          schema:
            type: string
        - name: tipe
          in: query
          description: Filter tipe penerima
          schema:
            type: string
        - name: status
          in: query
          description: Filter status pesan
          schema:
            type: integer
        - name: pesan
          in: query
          description: Pencarian kata kunci pada kolom pesan (LIKE)
          schema:
            type: string
        - name: replyToReseller
          in: query
          description: Checkbox filter reply ke reseller (kode_reseller IS NOT NULL)
          schema:
            type: boolean
        - name: perintahProvider
          in: query
          description: Checkbox filter perintah provider (is_perintah = 1)
          schema:
            type: boolean
        - name: cursor
          in: query
          description: Cursor pagination (kode record terakhir)
          schema:
            type: integer
            format: int64
      responses:
        <<: *errorResponses
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutboxResponse"
    post:
      summary: Insert Outbox
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      tags:
        - outbox
      parameters:
        - $ref: "#/components/parameters/TenantParam"
        - $ref: "#/components/parameters/DbSourceHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOutboxRequest"
      responses:
        <<: *errorResponses
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /api/v1/{tenant}/outbox/{kode}:
    put:
      summary: Update Outbox
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      tags:
        - outbox
      parameters:
        - $ref: "#/components/parameters/TenantParam"
        - $ref: "#/components/parameters/DbSourceHeader"
        - name: kode
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOutboxRequest"
      responses:
        <<: *errorResponses
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...